% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{add_communities}
\alias{add_communities}
\title{Append graph vertices with community assignment.}
\usage{
add_communities(x, communities, index_type = c("name", "index"))
}
\arguments{
\item{x}{an \code{igraph} object.}

\item{communities}{an object of class \code{communities}.}

\item{index_type}{type of the index used for assignment of the community
identifier attribute: names (default) of indexes.}
}
\value{
an \code{igraph} object.
}
\description{
Appends vertices of a graph with community identifiers.
}
\details{
The community identifiers are stored as a vertex attribute called
\code{community_id}. The community assignment may be re-named or re-coded by
calling \code{\link{comm_recode}} or \code{\link{comm_lump}}.
}
