% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{comm_lump}
\alias{comm_lump}
\title{Lump small communities together.}
\usage{
comm_lump(x, cutoff, other_name = "other")
}
\arguments{
\item{x}{a \code{communities} class object.}

\item{cutoff}{community size cutoff.}

\item{other_name}{name of the common category that groups all
small communities.}
}
\value{
a \code{communities} class object.
}
\description{
Collapses communities with size lower or equal to a cutoff into a common
category (e.g. 'other').
}
