% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{comm_recode}
\alias{comm_recode}
\title{Recode/rename community identifiers.}
\usage{
comm_recode(x, new_names)
}
\arguments{
\item{x}{an object of class \code{communities}.}

\item{new_names}{the re-coding vector as described in Details.}
}
\value{
an object of class \code{communities}.
}
\description{
Changes identifiers of communities with a vector. The identifiers are stored
as factors with the levels specified by names of the re-coding vector.
}
\details{
The re-coding character vector has a general form
\code{c(new_name_1 = old_name_1, new_name_2 = old_name_1, ...)}.
The community identifiers will be specified by a vector with levels
\code{c(new_name_1, new_name_2, ...)}.
}
