% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex_stats.R
\name{dimensions}
\alias{dimensions}
\title{Numbers of vertices and edges.}
\usage{
dimensions(x)
}
\arguments{
\item{x}{an \code{igraph} object.}
}
\value{
a numeric vector with vertex and edge number.
}
\description{
Extracts numbers of vertices and edges from an \code{igraph} object.
}
