% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_attributes.R
\name{get_vertex_attributes}
\alias{get_vertex_attributes}
\title{Retrieve vertex attributes.}
\usage{
get_vertex_attributes(x)
}
\arguments{
\item{x}{an \code{igraph} object.}
}
\value{
a tibble with vertex attributes.
}
\description{
Extracts all vertex attributes in a handy data frame. The first two columns
specify the vertex index and name.
}
