% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mtx_na_remove}
\alias{mtx_na_remove}
\title{Row/column-wise removal of observations from a symmetric matrix.}
\usage{
mtx_na_remove(x)
}
\arguments{
\item{x}{a symmetric matrix.}
}
\value{
a matrix.
}
\description{
Removes rows and columns of a symmetric matrix with at least one NA value.
}
