% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_modification.R
\name{neighbor_graph}
\alias{neighbor_graph}
\alias{neighbor_attr}
\title{Extraction of the neighborhood.}
\usage{
neighbor_graph(x, v = NULL, name = NULL)

neighbor_attr(x, v = NULL, name = NULL)
}
\arguments{
\item{x}{an \code{igraph} object.}

\item{v}{index of the vertex of interest or NULL. One of \code{v} or \code{name}
must be specified. If \code{v} is specified, \code{name} is ignored.}

\item{name}{name of the index of interest or NULL.  One of \code{v} or \code{name}
must be specified.}
}
\value{
\code{neighbor_graph()} returns an \code{igraph} class object,
\code{neighbor_attr()} returns a data frame with the neighbor indexes,
names (if specified), attributes, and weights of the edges between
the neighbors and the node of interest.
}
\description{
Functions \code{neighbor_graph()} and \code{neighbor_attr()} extract,
respectively, the neighborhood of a given vertex as a graph and attributes
of edges and nodes  for the neighbors of the given.
As such, the functions are handy supplements to
\code{\link[igraph]{neighbors}}.
}
