% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.igraph}
\alias{plot.igraph}
\title{Network visualization of a graph.}
\usage{
\method{plot}{igraph}(
  x,
  vertex_fill_variable = NULL,
  vertex_color_variable = vertex_fill_variable,
  vertex_shape_variable = NULL,
  weighting_order = 0,
  vertex_fill = "steelblue2",
  vertex_color = "steelblue4",
  vertex_alpha = 1,
  vertex_size = 2,
  edge_color = "gray60",
  edge_alpha = 1,
  edge_width = 0.25,
  label_vertices = FALSE,
  label_edges = FALSE,
  vertex_label_variable = "name",
  vertex_txt_color_variable = NULL,
  vertex_txt_size = 2.75,
  vertex_txt_color = "black",
  vertex_txt_face = "plain",
  vertex_labeller = identity,
  edge_txt_size = 2.75,
  edge_txt_color = "black",
  edge_txt_face = "plain",
  edge_txt_signif = 2,
  cust_theme = ggplot2::theme_void(),
  plot_title = NULL,
  plot_subtitle = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{an \code{igraph} object.}

\item{vertex_fill_variable}{name of an attribute coding for vertex symbol fill.}

\item{vertex_color_variable}{name of an attribute coding for vertex
symbol color.}

\item{vertex_shape_variable}{name of an attribute coding for vertex
symbol shape.}

\item{weighting_order}{order (power) of similarity weights between
the vertices, where similarity codes for edge alpha and width.
If \code{weighting_order = 0}, all edges will be plotted with the same width
and alpha.}

\item{vertex_fill}{fill of vertex symbols;
used only if \code{vertex_fill_variable = NULL}.}

\item{vertex_color}{color of vertex symbols;
used only if \code{vertex_color_variable = NULL}.}

\item{vertex_alpha}{alpha of vertex symbols.}

\item{vertex_size}{size of vertex symbols.}

\item{edge_color}{color of edges.}

\item{edge_alpha}{alpha of edges; used only if \code{weighting_order = 0}.}

\item{edge_width}{width of edges; used only if \code{weighting_order = 0}.}

\item{label_vertices}{logical: should labels of vertices be displayed?}

\item{label_edges}{logical; should labels with edge weights
(i.e. similarity) should be displayed?}

\item{vertex_label_variable}{name of an attribute storing the vertex labels;
by default this is the vertex name.}

\item{vertex_txt_color_variable}{name of an attribute coding for vertex
label text color.}

\item{vertex_txt_size}{size of vertex label text.}

\item{vertex_txt_color}{color of vertex label text; used only if
\code{vertex_txt_color_variable = NULL}.}

\item{vertex_txt_face}{font face of the vertex labels.}

\item{vertex_labeller}{a function that takes vertex labels and
transforms them.}

\item{edge_txt_size}{size of edge label text.}

\item{edge_txt_color}{color of edge label text.}

\item{edge_txt_face}{font face of edge label text.}

\item{edge_txt_signif}{significant digits used for rounding of the edge
weights presented in the edge labels.}

\item{cust_theme}{a custom \code{ggplot} theme.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot_subtitle. If \code{NULL}, numbers of nodes and edges
are displayed there.}

\item{seed}{random number generator seed. If \code{NULL}, no seed is set.}

\item{...}{additional arguments passed to
\code{\link[ggnetwork]{geom_nodetext_repel}}.}
}
\value{
a \code{ggplot} object.
}
\description{
Draws a network diagram of an \code{igraph} object using the tool set of
\code{ggnetwork} package. The plots may be easily customized by the user
by specifying color, fill, and alpha scales, as well as the plot theme
via functions of \code{ggplot2} package.
}
\details{
Technically, the \code{plot()} method for \code{igraph} objects uses the toolbox of
\code{ggnetwork} package.
}
