% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_modification.R
\name{prune_degree}
\alias{prune_degree}
\title{Delete low degree vertices.}
\usage{
prune_degree(x, cutoff = 0)
}
\arguments{
\item{x}{a \code{igraph} class object.}

\item{cutoff}{degree cutoff. Vertices with degree lower or equal to \code{cutoff}
will be removed.}
}
\value{
an \code{igraph} object.
}
\description{
The function \code{prune_degree()} removes vertices of the graph with degrees
equal or lower than a cutoff.
}
\details{
Internally, the function uses \code{\link[igraph]{delete_vertices}}.
}
