% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_modification.R
\name{prune_vertices}
\alias{prune_vertices}
\alias{select_vertices}
\title{Delete or select graph nodes with logical expression applied to attributes.}
\usage{
prune_vertices(x, ...)

select_vertices(x, ...)
}
\arguments{
\item{x}{a \code{igraph} class object.}

\item{...}{one or more logical expressions to select vertices
by their attributes.}
}
\value{
an \code{igraph} object.
}
\description{
The functions \code{prune_vertices()} and \code{select_vertices()} work in a manner
that resembles \code{\link[dplyr]{filter}} from \code{dplyr} package: the nodes
to be deleted (\code{prune_vertices()}) or kept (\code{select_vertices()}) are selected
by one or more logical operations on unquoted vertex attribute names.
}
