% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_attributes.R
\name{set_vertex_attributes}
\alias{set_vertex_attributes}
\title{Set vertex attributes.}
\usage{
set_vertex_attributes(x, lexicon, index_type = c("name", "index"))
}
\arguments{
\item{x}{an \code{igraph} object.}

\item{lexicon}{a data frame that contains vertex identifiers in the first
column and the vertex attributes to be specified in the remaining columns.
The new attributes will be named after the corresponding \code{lexicon} columns.}

\item{index_type}{type of the index used for assignment of the attributes:
names (default) of indexes.}
}
\value{
an \code{igraph} object.
}
\description{
Sets vertex attributes given a data frame by vertex names or indices.
}
