% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simil2graph}
\alias{simil2graph}
\title{Graph from a similarity matrix.}
\usage{
simil2graph(
  simil_mtx,
  cutoff = NULL,
  weighted = TRUE,
  diag = FALSE,
  na_action = c("ignore", "pad", "remove"),
  na_pad_value = 0,
  ...
)
}
\arguments{
\item{simil_mtx}{a symmetric similarity matrix.}

\item{cutoff}{a numeric value that defines the minimal similarity between
features. Only feature pairs with similarity larger or equal to \code{cutoff} will
become the graph's edges. If \code{NULL} (default), the all pairwise similarities
are used.}

\item{weighted}{a logical or character that specifies if and how the graph's
edges should be weighted by similarity between the features. Please consult
\code{\link[igraph]{graph_from_adjacency_matrix}} for details.}

\item{diag}{a logical that specifies if the diagonal of the similarity matrix
should be included the graph object. In most cases this results in
self-connecting edges. Defaults to \code{FALSE}.}

\item{na_action}{specifies how \code{NA} values are handled, see \strong{Details}.
Defaults to \code{"ignore"}.}

\item{na_pad_value}{a numeric value used to replace all \code{NA} values in the
similarity matrix. Used only when \code{na_action = "pad"}.}

\item{...}{extra arguments passed to \code{\link[igraph]{graph_from_adjacency_matrix}}}
}
\value{
a non-directional graph object of class \code{igraph}
}
\description{
Builds an \code{igraph} object at the top of a similarity matrix.
Internally, the function uses \code{\link[igraph]{graph_from_adjacency_matrix}}.
}
\details{
The \code{na_action} argument specifies, how \code{NA} values in the similarity matrix
are handled.
The default setting \code{na_action = "ignore"} determines that no action is taken,
and if there are any \code{NA} values the function stops and returns an error.
If \code{na_action = "pad"}, all \code{NA} values in the similatity matrix are replaced
with the numeric value provided with the \code{na_pad_value} argument.
Finally, if \code{na_action = "remove"}, rows and columns of the similarity
matrix with at least one \code{NA} are removed. '
}
