% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{sizes}
\alias{sizes}
\title{Size of communities.}
\usage{
sizes(communities)
}
\arguments{
\item{communities}{a \code{communities} class object.}
}
\value{
a \code{table}.
}
\description{
Extracts size of communities from a \code{communities} class object.
}
