% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_modification.R
\name{split_vertices}
\alias{split_vertices}
\title{Split a graph into subgraphs by levels of a vertex attribute.}
\usage{
split_vertices(x, ..., .drop = TRUE)
}
\arguments{
\item{x}{an \code{igraph} object.}

\item{...}{one or more unquoted attributes used for splitting.}

\item{.drop}{logical, should empty levels of the vector be skipped from the
output? Defaults to TRUE.}
}
\value{
a list of \code{igraph} objects.
}
\description{
The function splits a graph into subgraphs by factors of a vertex attribute
or a combination of vertex attributes.
}
