% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_objects.R
\name{as_iGraph}
\alias{as_iGraph}
\alias{as_iGraph.matrix}
\alias{as_iGraph.data.frame}
\alias{as_iGraph.dist}
\title{Create \code{igraph} objects.}
\usage{
as_iGraph(x, ...)

\method{as_iGraph}{matrix}(
  x,
  input_type = c("data", "similarity"),
  feature_type = c("columns", "rows"),
  fun = stats::cor,
  cutoff = NULL,
  weighted = TRUE,
  diag = FALSE,
  na_action = c("ignore", "pad", "remove"),
  na_pad_value = 0,
  ...
)

\method{as_iGraph}{data.frame}(
  x,
  feature_type = c("columns", "rows"),
  fun = stats::cor,
  cutoff = NULL,
  weighted = TRUE,
  diag = FALSE,
  na_action = c("ignore", "pad", "remove"),
  na_pad_value = 0,
  ...
)

\method{as_iGraph}{dist}(
  x,
  fun = function(x) x - 1,
  cutoff = NULL,
  weighted = TRUE,
  diag = FALSE,
  na_action = c("ignore", "pad", "remove"),
  na_pad_value = 0,
  ...
)
}
\arguments{
\item{x}{an object: a numeric matrix, numeric data frame or a \code{dist}
class object.}

\item{...}{extra arguments passed to the methods or to the \code{fun} function.}

\item{input_type}{specifies if a matrix should be treated as a data-storing
matrix (\code{input_type = 'data'}, default) or
a similarity matrix (\code{input_type = 'similarity'}).}

\item{feature_type}{specifies if features, i.e. vertices of the graph are
located in the rows or columns. Ignored for \code{dist} objects and
similarity matrices.}

\item{fun}{a function used to calculate similarity between features of a data
frame or matrix (\code{\link[stats]{cor}} by default) or to convert distances
to similarities (\code{1 - x} by default). Ignored for \code{dist} objects and
similarity matrices.}

\item{cutoff}{a numeric value that defines the minimal similarity between
features. Only feature pairs with similarity larger or equal to \code{cutoff} will
become the graph's edges. If \code{NULL} (default), the all pairwise similarities
are used.}

\item{weighted}{a logical or character that specifies if and how the graph's
edges should be weighted by similarity between the features. Please consult
\code{\link[igraph]{graph_from_adjacency_matrix}} for details.}

\item{diag}{a logical that specifies if the diagonal of the similarity matrix
should be included the graph object. In most cases this results in
self-connecting edges. Defaults to \code{FALSE}.
#' @param na_action specifies how \code{NA} values are handled, see \strong{Details}.
Defaults to \code{"ignore"}.}

\item{na_pad_value}{a numeric value used to replace all \code{NA} values in the
similarity matrix. Used only when \code{na_action = "pad"}.}
}
\value{
a non-directional graph object of class \code{igraph},
which may be processed, evaluated and visualized by tools of \code{igraph} package.
The methods \code{\link{summary.igraph}} and \code{plot()} which
allow for, respectively, computation of network node statistics and
visualization, are overwritten and allow for a richer output
and customization.
}
\description{
The generic function \code{as_iGraph()} takes a range of similarity, distance,
matrix, or data frame objects to generate \code{igraph} class graph objects based
on similarity or correlation between the features.
}
\details{
The function internally uses
\code{\link[igraph]{graph_from_adjacency_matrix}}.
The user can specify a similarity or correlation cutoff to select pairwise
associations of particular strength which will be used for the
graph construction.
If the function is provided with a numeric data frame or matrix, the
similarity between columns or rows will be calculated with an user-provided
function \code{fun} (\code{\link[stats]{cor}} by default).
For a matrix input, the user can specify if it should be treated as a
numeric, data-storing matrix or a similarity matrix.
If \code{dist} class objects are provided, they will be converted to similarity
matrices by a user-provided function \code{fun} (\code{1 - x} by default).
Any \code{NA} or negative values in the similarity matrix after application of the
similarity cutoff will raise warnings, any generation of the \code{igraph} object
may fail.
For matrices and data frames, row and column names are mandatory.

The \code{na_action} argument specifies, how \code{NA} values in the similarity matrix
are handled.
The default setting \code{na_action = "ignore"} determines that no action is taken,
and if there are any \code{NA} values the function stops and returns an error.
If \code{na_action = "pad"}, all \code{NA} values in the similatity matrix are replaced
with the numeric value provided with the \code{na_pad_value} argument.
Finally, if \code{na_action = "remove"}, rows and columns of the similarity
matrix with at least one \code{NA} are removed. '
}
