% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex_stats.R
\name{summary.igraph}
\alias{summary.igraph}
\title{Summary of vertex importance statistics.}
\usage{
\method{summary}{igraph}(object, ...)
}
\arguments{
\item{object}{an \code{igraph} object.}

\item{...}{extra arguments passed to methods, currently none.}
}
\value{
a tibble with statistics specified in Details.
}
\description{
Computes a bunch of node importance statistics for vertices of a graph.
}
\details{
The statistics are degree (number of edges projecting from a vertex),
betweenness (number of shortest paths between vertex pairs passing through
the vertex of interest), hub score (eigenvector of the similarity matrix).
}
